timetabler/requirements.txt
asgiref==3.2.10
Django==3.1.2
django-badgify==0.4.0
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.9.2
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
uWSGI==2.0.19.1
